fraudCertificate/requirements.txt
# Required packages for Certificate Authenticity Checker
flask==3.1.0
torch==2.8.0
torchvision==0.23.0
opencv-python==4.12.0.88
pillow==11.3.0
numpy==2.2.6
PyMuPDF==1.26.4
scikit-learn==1.5.2